+++ Rebus/Logging/IRebusLoggerFactory.cs
@@ -1,24 +1,20 @@
-using System.Diagnostics;
+using System;
 
 namespace Rebus.Logging
 {
     /// <summary>
-    /// Basic interface of a Rebus logger factory. If you intend to implement your own,
-    /// <see cref="AbstractRebusLoggerFactory"/> is the one to derive from - you should
-    /// probably not implement this interface directly.
+    /// Basic interface of a Rebus logger factory. You can make a tiny shortcut by deriving from <see cref="AbstractRebusLoggerFactory"/> if you intend to implement this interface
     /// </summary>
     public interface IRebusLoggerFactory
     {
         /// <summary>
-        /// Gets a logger that is initialized to somehow carry information on the class that is using it.
-        /// Be warned that this method will most likely be pretty slow, because it will probably rely on
-        /// some clunky <see cref="StackFrame"/> inspection.
-        /// </summary>
-        ILog GetCurrentClassLogger();
-
-        /// <summary>
         /// Gets a logger for the type <typeparamref name="T"/>
         /// </summary>
         ILog GetLogger<T>();
+
+        /// <summary>
+        /// Gets a logger for the type <paramref name="type"/>
+        /// </summary>
+        ILog GetLogger(Type type);
     }
 }