+++ Rebus/Pipeline/Receive/HandleDeferredMessagesStep.cs
@@ -46,7 +46,7 @@ This is done by checking if the incoming message has a '" + Headers.DeferredUnti
             _timeoutManager = timeoutManager;
             _transport = transport;
             _options = options;
-            _log = rebusLoggerFactory.GetCurrentClassLogger();
+            _log = rebusLoggerFactory.GetLogger<HandleDeferredMessagesStep>();
 
             var dueTimeoutsPollIntervalSeconds = (int)options.DueTimeoutsPollInterval.TotalSeconds;
             var intervalToUse = dueTimeoutsPollIntervalSeconds >= 1 ? dueTimeoutsPollIntervalSeconds : 1;