+++ Rebus/Testing/Internals/TestLoggerFactory.cs
@@ -12,7 +12,7 @@ namespace Rebus.Testing.Internals
 
         public IEnumerable<LogEvent> LogEvents => _logEvents.ToList();
 
-        protected override ILog GetLogger(Type type)
+        public override ILog GetLogger(Type type)
         {
             return new TestLogger(type, _logEvents);
         }