+++ CHANGELOG.md
@@ -1127,12 +1127,13 @@
 
 * Fix bug that can result in accidentally logging `IndexOutOfRangeExceptions`s when entering idle state, just when the server clock is adjusted back in time
 
-## 3.0.0-beta04
+## 3.0.0
 
 * Add synchronous bus API which can be accessed via `bus.Advanced.SyncBus` that exposes synchronous `bus.Send`, etc. operations
 * Add synchronous transaction context `DefaultSyncTransactionContext` which is completed with a synchronous `.Complete()` method
 * Move Rebus 1 (i.e. versions <= 0.84.0) legacy compatibility feature to separate package: [Rebus.LegacyCompatibility](https://github.com/rebus-org/Rebus.LegacyCompatibility) (no worries: Rebus 2 and 3 can talk just fine!)
 * Change transaction context API to be much more intuitive `DefaultTransactionContext` and `DefaultSyncTransactionContext` now automatically make the transaction context ambient
+* Removed `GetCurrentClassLogger` because the stack frame walking turned out to be unreliable in the face of what seems to be new, more aggressive method inlining
 
 
 ---